debug messages + uart + tuner + eprom
[debug messages]
Addded the support for message "tagging" in order to identify which
module generated the message. A proper #define must be added in top of
each module which needs this functionality, but it's just a matter of
copy and paste from the current implementations.

[uart]
during initialization, the UART also sends a "clear console" sequence
which clear the console on the monitoring computer. Of course this
simplify the discrimination of messages between consecutive sessions

[tuner]
a new module was added for supporting the most basic operations.
Unfortunately there's  no evidence that the peripheral is working
properly...

[eeprom]
also in this case a new module was introduced for adding basic
funcionalities for this peripheral. Even though it's extremely simple
for now, it supports eprom reading, resetting and programming. It has
also been tested and it works pretty good.